Removed namespace for zookeeper cluster (#140)
## Description

This PR removes the `namespace` field for the zookeeper reference; i.e. it is now not required to specify a namespace, instead the operator looks for the discovery config map in the same namespace as the druid cluster is.

This is the intended behaviour and also necessary for the kuttl tests to work.

## Review Checklist
- [x] Code contains useful comments
- [x] (Integration-)Test cases added (or not applicable)
- [x] Documentation added (or not applicable)
- [x] Changelog updated (or not applicable)
- [x] Cargo.toml only contains references to git tags (not specific commits or branches)
- [x] Helm chart can be installed and deployed operator works (or not applicable)

Once the review is done, comment `bors r+` (or `bors merge`) to merge. [Further information](https://bors.tech/documentation/getting-started/#reviewing-pull-requests)


Co-authored-by: Stacky McStackface <stackable-bot@users.noreply.github.com>