Feature/s3 extension defaults (#192)
## Description

It is not immediately evident from the Apache Druid documentation that only one s3-endpoint can be used per druid cluster: the endpoint cannot be entered in the console UI and must be supplied as a runtime property (only buckets/baseKeys can be entered in the console). The credentials can be passed either as environment variables, as runtime properties, or directly via the UI, so they do not need to exist as runtime properties necessarily.

The upshot of this is that the druid-s3-extension requires an s3-endpoint for it to be initialized: making the loading of the s3-extension conditional upon the provision of an `s3:endpoint` in our CR means that s3-functionality will only be activated in the console if it can be used (i.e. that an endpoint is also available).

The following have been tested:

- quickstart: include s3-extension in list and set a value for `druid.s3.endpoint.url`
  - ingest sample- and s3-data (adding credentials/bucket in UI)
- operator + hdfs-deep-storage + without s3 definition in CR: ingestion of sample data
- operator + hdfs-deep-storage + with s3 definition in CR: ingestion of s3-data
- operator + s3-deep-storage: ingestion of s3 and non-s3 data
- existing integration tests