Getting started docs (#278)
# Description

https://github.com/stackabletech/documentation/issues/237

This is a first "getting started" section for an operator, which we hopefully can bring to other operators afterwards.

This replaces the old install page, and takes parts of the usage page and the examples to build it all into a "Getting started" guide/tutorial.

Also, all the script snippets are in one big script which can be executed and functions as a test.

Also, templating was introduced to be able to template operator versions into the script. I wanted to avoid this, but didn't see an alterantive way to get Helm and stackablectl to install the correct versions. I think with the templating there's just one file that needs to be adapted, which we can also automate in the future.

I've removed the simple-druid-cluster example, because that's now in the docs (and with the script it is also testable in a semi-automatic way: we can fully automate this in the future).

If this PR gets accepted, i'd like to put the `docs_templating.sh` script into operator-templating.



Co-authored-by: Felix Hennig <fhennig@users.noreply.github.com>