Fix OOM error with manual buffer size specification (#380)
# Description

fixes #359 

automatic buffer sizing seems to be the problem in #359 

This PR requires a new operator-rs version with the operator-rs features from this PR https://github.com/stackabletech/operator-rs/pull/544

This PR is intended to fix this, by calculating a buffer size based on the user provided memory request. Druid guidelines are used to compute all values.
We also set the available `MaxDirectAccessMemory` dynamically, instead of hardcoding it. The heap calculation now incorporates this value, so now it shouldn't happen anymore that the JVM tries to allocate more than there is available to the container.

The computation has a few other benefits:
- maximum buffer size can now be 2GB instead of 1GB (because we're not using druid `auto` anymore)
- a fix memory quantity is reserved for the OS, not a scaled one. This saves memory if more memory is available
- there we're bugs that didn't surface yet with the MaxDirectAccessMemory of all roles. All of them are fixed now, the MaxDirectAccessMemory + Heap Memory does not exceed the maximum memory available anymore.
- The rounding used for the JVM memory is not unit-dependent anymore. Before it would round of .2 of _whatever_ quantity, so there might be 200MB sitting around unused. This is not the case anymore.
- If the direct access memory is maxed out, all the remaining memory will be allocated as Heap, to make maxium use of the allocatable memory. This is also for all roles.